fix: clippy warnings, common dependencies

+++ halo2_frontend/src/circuit.rs
@@ -88,7 +88,7 @@ pub fn compile_circuit<F: Field, ConcreteCircuit: Circuit<F>>(
         circuit.params(),
     );
 
-    if (n as usize) < cs.minimum_rows() {
+    if n < cs.minimum_rows() {
         return Err(Error::not_enough_rows_available(k));
     }
     let mut assembly = plonk::keygen::Assembly {
